Integrate WSGI code into `http_server.py`
